* dialign/prob.c (calc_score_dist): Fix compiler warnings. * dialign/io.c (read_anchors): Ditto. * dialign/translate.c (translate, inverse, retranslate): Ditto.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@67131 ad02020f-2a26-4133-947e-1829317c1416

+++ dialign/io.c
@@ -525,7 +525,7 @@ struct simple_diag_col* read_anchors(char *filename, struct seq_col* scol) {
 
   //fscanf( fp, "%li %li %Le\n", &ti,&tscr,&weight );
 
-  int s1,s2,sp1,sp2,len;
+  long int s1,s2,sp1,sp2,len;
   double score;
   
   int alloc_size = 64;
